refactor: consolidate contract files and remove redundant barrels
