--- src/lib/profiles/host.repository.contract.ts
@@ -1,17 +1,0 @@
-import { Networks } from "@/lib/mainsail";
-
-export type Host = Networks.NetworkHost;
-export type HostSet = { name: string; host: Host }[];
-export type HostMap = Record<string, HostSet>;
-
-export interface IHostRepository {
-	all(): Record<string, HostMap>;
-
-	allByNetwork(network: string): HostSet;
-
-	push(data: { host: Host; name: string; network: string }): HostSet;
-
-	fill(entries: object): void;
-
-	forget(network: string, index?: number): void;
-}
