--- src/lib/profiles/network.repository.contract.ts
@@ -1,18 +1,0 @@
-import { Networks } from "@/lib/mainsail";
-
-export type Network = Networks.NetworkManifest;
-export type NetworkMap = Record<string, Network>;
-
-export interface INetworkRepository {
-	all(): NetworkMap;
-
-	allByCoin(coin: string): Network[];
-
-	get(network: string): Network;
-
-	push(host: Network): Network;
-
-	fill(entries: object): void;
-
-	forget(network: string): void;
-}
