+++ src/lib/profiles/profile.contract.ts
@@ -11,8 +11,7 @@ import {
 	IWalletRepository,
 } from "./contracts.js";
 import { AttributeBag } from "./helpers/attribute-bag.js";
-import { IHostRepository } from "./host.repository.contract.js";
-import { INetworkRepository } from "./network.repository.contract.js";
+import { IHostRepository, INetworkRepository } from "./contracts.js";
 import { LedgerService } from "@/lib/mainsail/ledger.service.js";
 import { ValidatorService } from "./validator.service.js";
 import { ExchangeRateService } from "./exchange-rate.service.js";
@@ -104,4 +103,8 @@ export interface IProfile {
 	whitelistedContractAddresses(): string[];
 	whitelistContractAddress(address: string): string[];
 	removeWhitelistedContractAddress(address: string): string[];
+}
+
+export interface IProfileExporter {
+	export(password?: string, options?: IProfileExportOptions): Promise<string>;
 }
