--- src/lib/profiles/profile.exporter.contract.ts
@@ -1,13 +1,0 @@
-import { IProfileExportOptions } from "./contracts.js";
-
-export interface IProfileExporter {
-	/**
-	 * Export the profile data to a base64 string.
-	 *
-	 * @param {string} [password]
-	 * @param {IProfileExportOptions} [options]
-	 * @return {string}
-	 * @memberof Profile
-	 */
-	export(password?: string, options?: IProfileExportOptions): Promise<string>;
-}
