+++ src/lib/profiles/profile.ts
@@ -17,7 +17,7 @@ import {
 import { DataRepository } from "./data.repository";
 import { AttributeBag } from "./helpers/attribute-bag";
 import { Avatar } from "./helpers/avatar";
-import { IHostRepository } from "./host.repository.contract";
+import { IHostRepository } from "./contracts";
 import { HostRepository } from "./host.repository";
 import { NetworkRepository } from "./network.repository";
 import { PasswordManager } from "./password";
