--- src/lib/profiles/repositories.ts
@@ -1,1 +1,0 @@
-export * from "./data.repository";
